וגם את שלך

+++ app/src/main/java/com/example/davidsalmon/android5778_8694_7928_01/model/backend/DB_manager.java
@@ -14,7 +14,7 @@ public interface DB_manager {
 
     boolean UserExistsOnDataBase(Long ID);
     long addUser(ContentValues newClient);
-    long addModel(ContentValues newModel);
+    int addModel(ContentValues newModel);
     String addCar(ContentValues newCar);
     int addBranch(ContentValues newBranch);
     List<CarsModel> AllCarsModel();
